requirements.txt
flask
flask_socketio
flask-login 
PySocks
python-dotenv